Avoided non-deletion of too new directories
In other words, this change restores the old behaviour for directories.
Apparently, changes within the directory may appear it too new, which prevents
deletion, although that deletion should happen.

This makes this feature not any more behaving as documented.  Thus, it cannot
be distributed this way.  Instead, in the long run, we must detect the true
timestamp of each directory on the destination before it is modified.